+++ map_solomon.py
